% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyPlot-package.R
\docType{package}
\name{polyPlot-package}
\alias{polyPlot-package}
\title{polyPlot: A Tool for Visualizing Distributional Statistics}
\description{
polyPlot is a small R package that offers a modified version of Seier and Bonett's (2011) polyplot,
a visualization tool for various distributional summary statistics.
}
\details{
The polyPlot package provides functions to create enhanced polyplots, which visually summarize key
distributional statistics such as means, variances, quantiles, and confidence intervals.
These plots facilitate intuitive understanding of data distributions and are useful in exploratory data analysis
and teaching statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwendorf/polyPlot}
  \item Report bugs at \url{https://github.com/cwendorf/polyPlot/issues}
}

}
\author{
\strong{Maintainer}: Craig Wendorf
\email{cwendorf@uwsp.edu}
}
