% File man/polyPlot2-package.Rd

\name{polyPlot2-package}
\alias{polyPlot2-package}
\docType{package}

\title{polyPlot2}
\description{
\code{polyPlot2} A modified polyplot.
}

\details{
\tabular{ll}{
Package: \tab polyPlot2\cr
Type: \tab Package\cr
}
A tool for visualizing distributional statistics.
}

\author{
Craig A. Wendorf\cr 
Department of Psychology\cr 
University of Wisconsin - Stevens Point\cr 
\email{cwendorf@uwsp.edu} 
}