% File man/polyStats.Rd
% Part of the polyplot2 package

\name{polyStats}
\alias{polyStats}
\title{polyStats}
\description{Provides PolyPlot statistics}

\usage{
\method{polyStats}{default}(x,digits=2)
}

\arguments{
\item{x}{a (non-empty) numeric vector of data values}
\item{digits}{number of decimal places for the statistics}
}