% File man/polyPlot-package.Rd

\name{polyPlot-package}
\alias{polyPlot-package}
\docType{package}

\title{polyPlot}
\description{
\code{polyPlot} A modified polyplot.
}

\details{
\tabular{ll}{
Package: \tab polyPlot\cr
Type: \tab Package\cr
}
A tool for visualizing distributional statistics.
}

\author{
Craig A. Wendorf\cr 
Department of Psychology\cr 
University of Wisconsin - Stevens Point\cr 
\email{cwendorf@uwsp.edu} 
}