% File man/polyplot2-package.Rd

\name{polyplot2-package}
\alias{polyplot2-package}
\alias{polyplot2}
\docType{package}

\title{polyplot2}
\description{
\code{polyplot2} A modified polyplot.
}

\details{
\tabular{ll}{
Package: \tab polyplot2\cr
Type: \tab Package\cr
}
A modified polyplot for R.
}

\author{
Craig A. Wendorf\cr 
Department of Psychology\cr 
University of Wisconsin - Stevens Point\cr 
\email{cwendorf@uwsp.edu} 
}